Start / End of Turn
Since several rulings state there are moments at the start and / or end of turn - in difference to the Unlock / Discard phases - these two timing windows should be made explicit. 
While these windows might not need to be explicit under consideration of the Golden Rule of Cards, having them stated does make it easier for new players to distinguish when effects take place.