Restructrure
Added explanations which were in the glossaries, moved fundamental rules into their own section (along with "important terms of the game"). 
With the latter change, it might be easier to read through the explanations of what the card types are. The change might also reduce the possibility to overlook important definitions due to their increased prominence.