Suggestion
Suggestion to slighly simplify Dodge: Remove that Retainers are not protected. So dodge protects not only equipment, but all possessions.

Reasoning: It seems like a very corner case, given the special setup which needs to be met for retainers to be targeted. It also is an exception to the rule which does not provide any real tactical depth and is hard to memorize.