Remove the option to wake in the "as played" window.
This is a rule change.

Given the new modifier cards, it seems mandatory to remove this old exception. It only impacts Power of All and Rewind Time, which should not be an issue. On the other hand, it makes Unleashing the Bestial Soul more obvious.

Right now, rulemongers can use this exception to play a wake after Unleashing the Bestial Soul has been played but before its effect is enacted, in the "as played" window. It's very unnatural and unexpected for most players. I had to explain this multiple times even to veterans until they understood what was possible here.