Trifle action
Suggested **rules change**: 

Adding trifle actions so Trifles can be played using them. That removes the necessity for trifles to give another master phase action, and enables a play of a trifle after a regular master card. 

The suggested changes falls in line with how other "does not count against the limit (strikes / bleed)" are handled. Out-of-turn-Masters are suggested to cost both a trifle and master phase action to prevent a change of power.

The change is also simpler to learn and could enhance game tempo a bit..