The original oust language uses "controlled" with a different meaning than the rest of the rulebook.  When ousted it was using controlled to mean everything in your "play area," but controlled _has_ a specific meaning in VTES.  The language here must reflect that because RAW would mean that an ousted Methuselah's hand, library, unctronlled region, ash heap, and contested cards would not be removed from the game because, by definition, none of those are "controlled."
The cases that need to be solved to be accurate with what we want:
- The ousted Methulah's hand, ash heap, library, and uncontrolled region are removed
- All contested cards are immediately yielded
- All controlled cards are removed

I am not thrilled by the long sentence that I made to describe that.  It was initially shorter, but I found additional cases as I was writing.  Perhaps a bulleted list would be better?

Regardless, the use of "controlled" in the original text is misleading at best and RAW a terrible description of the intent.